# Optimise offline session exporter